refactor: move incrementPlotContentAge to its own file
